dev同步master

Created-by: mazhixin00_00
Commit-by: weixin_44144262;LagavulinNeat;guowenna1;zhou-wenxue;hyh_hh;changetheway;ascend-robot;lanwangli;mazhixin00_00;chensiyu47;huyinghong 00566667;TwistZzzz;hejianping
Merged-by: lanwangli
Description: dev同步master

See merge request: Ascend/MindIE-SD!223

+++ mindiesd/layers/flash_attn/ascend_laser_attention.py
@@ -70,7 +70,7 @@ class AscendLaserAttention(AttentionOperateBase):
     @staticmethod
     def la_postprocess_output(attention_out, dtype, qseqlen, head_dim):
         if dtype != attention_out.dtype:
-            attention_out = attention_out.to(dtype)
+            attention_out = attention_out.to(torch.float16).to(dtype)
         attention_out = attention_out[:, :, :qseqlen, :head_dim]
         return attention_out
